refactor/#47: UserDocumentLastOpened를 user domain에 속하도록 변경
