+++ src/main/java/com/haru/api/global/aspect/LastOpenedAspect.java
@@ -2,7 +2,7 @@ package com.haru.api.global.aspect;
 
 import com.haru.api.global.common.Documentable;
 import com.haru.api.user.domain.UserDocumentId;
-import com.haru.api.workspace.application.port.in.UserDocumentLastOpenedQueryUseCase;
+import com.haru.api.user.application.port.in.UserDocumentLastOpenedQueryUseCase;
 import com.haru.api.user.domain.User;
 import com.haru.api.global.annotation.TrackLastOpened;
 import lombok.RequiredArgsConstructor;
